Add files via upload
the creation of a boxplot and the use of a seed to ensure the values are the same so when describing the aspects and plots of a box plot it remains the same no matter how many times the code is ran. The median, first and third quartile are discussed along with the outliers and the upper and lower limits.